backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Pedro Gustavo Santana <pedrogustavosantan97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pdfminer-six = "^20231228"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
